requirements.txt
requests >= 2.32.5
python-dotenv >= 1.1.1
pandas >= 2.3.2
numpy >= 2.3.3
jupyterlab >= 4.4.7
matplotlib >= 3.10.6
seaborn >= 0.13.2
plotly >= 6.3.0
